chore: separate application code from server code
